refactor: displace scroll-to logic in a hook
